last_in_3_sem/requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-postgres
apache-airflow-providers-mongo
protobuf<5
pymongo
psycopg2-binary
pandas
faker
python-dotenv
